@@ -15,17 +15,7 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib,vendor}/**/*"] + ["MIT-LICENSE", "Rakefile", "README.md", "deprecations.rb"]
 
-<<<<<<< HEAD
-  s.add_dependency "rails", ">= 3.1.3"
-  s.add_dependency "bcrypt-ruby"
-  s.add_dependency "coffee-rails"
-  s.add_dependency "sass-rails",'~> 3.2.3'
-  s.add_dependency "truncate_html"
-  s.add_dependency "jquery-rails"
-  s.add_dependency "ckeditor"
-  s.add_dependency 'select2-rails'
-=======
-  s.add_dependency "rails", "= 4.2.7"
+  s.add_dependency "rails", "> 5.0.7"
   s.add_dependency "bcrypt", '~> 3.1.7'
   s.add_dependency "coffee-rails",'>= 4.0.0'
   s.add_dependency "truncate_html"
@@ -35,7 +25,6 @@ Gem::Specification.new do |s|
   s.add_dependency 'select2-rails', '~> 3.2'
   s.add_dependency "sass-rails",'~> 5.0.0'
   s.add_dependency "responders", '~> 2.1.0'
->>>>>>> jipiboily/master
 
   s.add_development_dependency "rspec-rails", "~> 2.8"
   s.add_development_dependency 'factory_girl_rails', '~> 1.4.0'
