@@ -1,5 +1,5 @@
 class Monologue::Admin::SessionsController < Monologue::Admin::BaseController
-  skip_before_filter :authenticate_user!
+  skip_before_action :authenticate_user!
   
   def new
   end
